add gradient test case
