+++ ceviche_challenges/beam_splitter/jax_model.py
@@ -1,0 +1,3 @@
+from ._model import generateModel
+
+BeamSplitterModel = generateModel(True)
