+++ ceviche_challenges/mode_converter/jax_model.py
@@ -1,0 +1,3 @@
+from ._model import generateModel
+
+ModeConverterModel = generateModel(True)
