+++ ceviche_challenges/waveguide_bend/jax_model.py
@@ -1,0 +1,3 @@
+from ._model import generateModel
+
+WaveguideBendModel = generateModel(True)
