+++ ceviche_challenges/wdm/__init__.py
@@ -13,5 +13,6 @@
 # limitations under the License.
 
 import ceviche_challenges.wdm.model
+import ceviche_challenges.wdm.jax_model
 import ceviche_challenges.wdm.prefabs
 import ceviche_challenges.wdm.spec
